requirements.txt
flask==3.0.0
flask-socketio==5.3.6
pymongo==4.6.1
python-dotenv==1.0.1
redis==5.0.1
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
pandas==2.2.0
numpy==1.26.3
gunicorn==21.2.0
eventlet==0.35.2 